pyproject.toml
[project]
name = "intercompany-dispute-env"
version = "0.1.0"
description = "Multi-agent financial dispute resolution environment for OpenEnv"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openenv-core>=0.2.2",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "fastmcp>=2.0.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "huggingface-hub>=1.8.0",
]

[project.scripts]
server = "server.app:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["server*", "domain*", "services*", "graders*", "tasks*", "agent*"]

[dependency-groups]
dev = [
    "huggingface-hub>=1.8.0",
    "pytest>=9.0.2",
]
